Refactored to es6 modules and replaced jshint with eslint
