+++ .babelrc
@@ -1,4 +1,3 @@
 {
-  "presets": ["@babel/preset-env"],
-  "plugins": ["@babel/plugin-syntax-dynamic-import"]
+  "presets": ["@babel/preset-env"]
 }
