+++ demos/serialization.html
@@ -16,7 +16,7 @@
     <script type="text/javascript" src="assets/rainbow-custom.min.js"></script>
 
     <script type="text/javascript" src="assets/ColorPicker.js"></script>
-    <script type="text/javascript" src="../src/TextHighlighter.js"></script>
+    <script type="text/javascript" src="../build/TextHighlighter.js"></script>
   </head>
   <body>
     <div class="view">
