+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "TextHighlighter",
+  "name": "texthighlighter",
   "version": "1.2.0",
   "description": "TextHighlighter",
   "dependencies": {},
@@ -8,16 +8,28 @@
     "@babel/core": "^7.5.5",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/preset-env": "^7.5.5",
+    "babel-eslint": "^10.0.3",
     "babel-jest": "^24.9.0",
+    "babelify": "^10.0.0",
+    "eslint": "^6.2.2",
+    "eslint-config-prettier": "^6.1.0",
+    "eslint-plugin-jest": "^22.15.2",
+    "eslint-plugin-prettier": "^3.1.0",
     "express": "^4.17.1",
     "grunt": "^0.4.5",
+    "grunt-babel": "^8.0.0",
+    "grunt-browserify": "^5.3.0",
     "grunt-closurecompiler": "^0.9.9",
     "grunt-contrib-clean": "^0.6.0",
     "grunt-contrib-jshint": "^0.10.0",
+    "grunt-contrib-uglify": "^4.0.1",
+    "grunt-contrib-watch": "^1.1.0",
     "grunt-jsdoc": "^0.5.7",
     "jest": "^24.9.0",
     "jest-extended": "^0.11.2",
-    "jsdom": "^15.1.1"
+    "jsdom": "^15.1.1",
+    "load-grunt-tasks": "^5.1.0",
+    "prettier": "^1.18.2"
   },
   "jest": {
     "testURL": "https://localhost",
@@ -46,6 +58,7 @@
     "type": "git",
     "url": "https://github.com/mir3z/texthighlighter.git"
   },
+  "main": "index.js",
   "author": "mirz <mirz.hq@gmail.com>",
   "license": "MIT"
 }
