+++ src/global-script.js
@@ -1,0 +1,7 @@
+import TextHighlighter from "./text-highlighter";
+
+/**
+ * Expose the TextHighlighter class globally to be
+ * used in demos and to be injected directly into html files.
+ */
+global.TextHighlighter = TextHighlighter;
