+++ src/index.js
@@ -1,0 +1,3 @@
+import TextHighlighter from "text-highlighter";
+
+export default TextHighlighter;
